+++ multiversx_sdk/controllers/relayed_controller.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from multiversx_sdk.core.controllers.interfaces import IAccount
+from multiversx_sdk.controllers.interfaces import IAccount
 from multiversx_sdk.core.interfaces import ITransaction
 from multiversx_sdk.core.transaction import Transaction
 from multiversx_sdk.core.transaction_computer import TransactionComputer
