+++ multiversx_sdk/controllers/smart_contract_controller.py
@@ -2,12 +2,10 @@ from pathlib import Path
 from typing import Any, List, Optional, Protocol, Sequence, Union
 
 from multiversx_sdk.adapters.query_runner_adapter import QueryRunnerAdapter
+from multiversx_sdk.controllers.network_provider_wrapper import ProviderWrapper
+from multiversx_sdk.controllers.token_management_controller import IAccount
 from multiversx_sdk.converters.transactions_converter import \
     TransactionsConverter
-from multiversx_sdk.core.controllers.network_provider_wrapper import \
-    ProviderWrapper
-from multiversx_sdk.core.controllers.token_management_controller import \
-    IAccount
 from multiversx_sdk.core.interfaces import IAddress
 from multiversx_sdk.core.smart_contract_queries_controller import \
     SmartContractQueriesController
