+++ multiversx_sdk/controllers/token_management_controller.py
@@ -1,9 +1,8 @@
 from typing import List, Protocol
 
+from multiversx_sdk.controllers.network_provider_wrapper import ProviderWrapper
 from multiversx_sdk.converters.transactions_converter import \
     TransactionsConverter
-from multiversx_sdk.core.controllers.network_provider_wrapper import \
-    ProviderWrapper
 from multiversx_sdk.core.interfaces import IAddress
 from multiversx_sdk.core.transaction import Transaction
 from multiversx_sdk.core.transaction_computer import TransactionComputer
