+++ multiversx_sdk/facades/__init__.py
@@ -1,0 +1,9 @@
+from multiversx_sdk.facades.account import Account
+from multiversx_sdk.facades.entrypoints import (DevnetEntrypoint,
+                                                MainnetEntrypoint,
+                                                NetworkEntrypoint,
+                                                TestnetEntrypoint)
+
+__all__ = [
+    "Account", "DevnetEntrypoint", "MainnetEntrypoint", "NetworkEntrypoint", "TestnetEntrypoint"
+]
